refactoring the authentication as a custom hook
